refactor(currencies): drop the @ledgerhq/types-cryptoassets package
