+++ e2e/mobile/package.json
@@ -40,7 +40,6 @@
     "@ledgerhq/live-e2e-shared": "workspace:^",
     "@ledgerhq/live-wallet": "workspace:^",
     "@ledgerhq/test-utils": "workspace:^",
-    "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-devices": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
     "@types/body-parser": "1.19.5",
