+++ libs/coin-modules/coin-aptos/.unimportedrc.json
@@ -2,8 +2,7 @@
   "ignoreUnresolved": [
     "@jest/expect-utils",
     "@jest/get-type",
-    "jest-util",
-    "@ledgerhq/types-cryptoassets"
+    "jest-util"
   ],
   "entry": [
     "src/index.ts",
@@ -14,7 +13,6 @@
   ],
   "ignoreUnused": [
     "@ledgerhq/devices",
-    "@ledgerhq/types-cryptoassets",
     "invariant",
     "got",
     "@ledgerhq/errors"
