+++ libs/coin-modules/coin-casper/.unimportedrc.json
@@ -20,8 +20,7 @@
     "jest-util",
     "jest-get-type",
     "jest-matcher-utils",
-    "jest-message-util",
-    "@ledgerhq/types-cryptoassets"
+    "jest-message-util"
   ],
   "ignoreUnused": [
     "@ledgerhq/errors"
