+++ libs/coin-modules/coin-celo/.unimportedrc.json
@@ -19,8 +19,7 @@
     "jest-util",
     "jest-get-type",
     "jest-matcher-utils",
-    "jest-message-util",
-    "@ledgerhq/types-cryptoassets"
+    "jest-message-util"
   ],
   "ignoreUnimported": [
     "src/bridge/transaction.ts",
