+++ libs/coin-modules/coin-evm/.unimportedrc.json
@@ -36,7 +36,6 @@
     "@ledgerhq/cryptoassets-evm-signatures/data/evm/index",
     "bn.js",
     "lru-cache",
-    "@ledgerhq/types-cryptoassets",
     "@ledgerhq/types-live"
   ],
   "ignoreUnused": [
