+++ libs/coin-modules/coin-sui/.unimportedrc.json
@@ -32,7 +32,6 @@
   ],
   "ignoreUnused": [
     "@ledgerhq/live-env",
-    "@ledgerhq/types-cryptoassets",
     "@ledgerhq/errors"
   ]
 }
