+++ libs/coin-modules/coin-ton/.unimportedrc.json
@@ -20,7 +20,6 @@
   ],
   "ignoreUnused": [
     "msw",
-    "@ledgerhq/types-cryptoassets",
     "@ledgerhq/live-env",
     "@ledgerhq/errors"
   ]
