+++ libs/ledger-live-common/.unimportedrc.json
@@ -1023,7 +1023,6 @@
     "superstruct",
     "use-sync-external-store/shim",
     "whatwg-fetch",
-    "@ledgerhq/types-cryptoassets",
     "@ledgerhq/types-live",
     "@ledgerhq/types-devices"
   ]
