--- libs/ledgerjs/packages/types-cryptoassets/.unimportedrc.json
@@ -1,3 +1,0 @@
-{
-  "entry": ["src/index.ts"]
-}
