--- libs/ledgerjs/packages/types-cryptoassets/README.md
@@ -1,8 +1,0 @@
-<img src="https://user-images.githubusercontent.com/4631227/191834116-59cf590e-25cc-4956-ae5c-812ea464f324.png" height="100" />
-
-## @ledgerhq/types-cryptoassets
-
-> [!NOTE]
-> **Status: STABLE** — Production-ready; API is considered stable.
-
-Ledger types for crypto assets and tokens.
