--- libs/ledgerjs/packages/types-cryptoassets/jest.config.ts
@@ -1,10 +1,0 @@
-import { dirname } from "node:path";
-import { fileURLToPath } from "node:url";
-import baseConfig from "../../jest.config.ts";
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-export default {
-  ...baseConfig,
-  rootDir: __dirname,
-};
