--- libs/ledgerjs/packages/types-cryptoassets/package.json
@@ -1,62 +1,0 @@
-{
-  "name": "@ledgerhq/types-cryptoassets",
-  "version": "7.39.0",
-  "description": "Ledger types for crypto assets and tokens",
-  "keywords": [
-    "Ledger"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/LedgerHQ/ledger-live.git"
-  },
-  "bugs": {
-    "url": "https://github.com/LedgerHQ/ledger-live/issues"
-  },
-  "homepage": "https://github.com/LedgerHQ/ledger-live/tree/develop/libs/ledgerjs/packages/types-cryptoassets",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "lib/index.js",
-  "module": "lib-es/index.js",
-  "types": "lib/index.d.ts",
-  "license": "Apache-2.0",
-  "scripts": {
-    "clean": "rimraf lib lib-es",
-    "build": "tsc --project tsconfig.build.json && tsc --project tsconfig.build.json -m esnext --moduleResolution bundler --outDir lib-es",
-    "prewatch": "pnpm build",
-    "watch": "tsc --watch --project tsconfig.build.json",
-    "watch:es": "tsc --watch --project tsconfig.build.json -m esnext --moduleResolution bundler --outDir lib-es",
-    "lint": "oxlint ./src",
-    "lint:fix": "pnpm lint --fix",
-    "test": "jest",
-    "coverage": "jest --coverage",
-    "unimported": "unimported"
-  },
-  "devDependencies": {
-    "@types/jest": "catalog:",
-    "@types/node": "catalog:",
-    "jest": "catalog:",
-    "rimraf": "^4.4.1",
-    "source-map-support": "^0.5.21",
-    "@swc/jest": "catalog:",
-    "@swc/core": "catalog:",
-    "ts-node": "^10.4.0"
-  },
-  "exports": {
-    ".": {
-      "@ledgerhq/source": "./src/index.ts",
-      "import": "./lib-es/index.js",
-      "require": "./lib/index.js",
-      "default": "./lib/index.js"
-    },
-    "./lib-es/*": "./lib-es/*.js",
-    "./lib/*": "./lib/*.js",
-    "./*": {
-      "@ledgerhq/source": "./src/*.ts",
-      "import": "./lib-es/*.js",
-      "require": "./lib/*.js",
-      "default": "./lib/*.js"
-    },
-    "./package.json": "./package.json"
-  }
-}
