--- libs/ledgerjs/packages/types-cryptoassets/src/slip44.ts
@@ -1,104 +1,0 @@
-// As defined here: https://github.com/satoshilabs/slips/blob/master/slip-0044.md
-export enum CoinType {
-  AE = 457,
-  AION = 425,
-  AKA = 200625,
-  ALEO = 683,
-  ALGO = 283,
-  APTOS = 637,
-  ATOM = 118,
-  ARK = 111,
-  ATH = 1620,
-  BANANO = 198,
-  BITTENSOR = 1005,
-  BTC = 0,
-  BTC_CASH = 145,
-  BTC_GOLD = 156,
-  BTC_PRIVATE = 183,
-  BTC_TESTNET = 1,
-  CALLISTO = 820,
-  CANTON_NETWORK = 6767,
-  CARDANO = 1815,
-  CELO = 52752,
-  CONCORDIUM = 919,
-  CRYPTO_ORG = 394,
-  CSPR = 506,
-  DASH = 5,
-  DECRED = 42,
-  DEXON = 237,
-  DIGIBYTE = 20,
-  DOGE = 3,
-  ELASTOS = 2305,
-  ELLAISM = 163,
-  MULTIVERSX = 508,
-  ENERGY_WEB_CHAIN = 246,
-  EOS = 194,
-  ETH = 60,
-  ETH_CLASSIC = 61,
-  EXPANSE = 40,
-  FACTOM = 131,
-  FIC = 5248,
-  FILECOIN = 461,
-  FLOW = 539,
-  GAME = 101,
-  GOCHAIN = 6060,
-  GRS = 17,
-  HEDERA = 3030,
-  HELIUM = 904,
-  HPB = 269,
-  HYCON = 1397,
-  ICON = 4801368,
-  ICP = 223,
-  IOTA = 4218,
-  IOV = 234,
-  KASPA = 111111,
-  KIN = 2017,
-  KOMODO = 141,
-  KUSAMA = 434,
-  LBRY = 140,
-  LITECOIN = 2,
-  LISK = 134,
-  MINA = 12586,
-  MIX = 76,
-  MONERO = 128,
-  MUSICOIN = 184,
-  NANO = 165,
-  NEAR = 397,
-  NEM = 43,
-  NEO = 888,
-  NERVOS = 309,
-  NIMIQ = 242,
-  NIX = 400,
-  NOS = 229,
-  ONTOLOGY = 1024,
-  PARTICL = 44,
-  PIRL = 164,
-  POA = 178,
-  POLKADOT = 354,
-  QRL = 238,
-  QTUM = 88,
-  RAVECOIN = 175,
-  RIPPLE = 144,
-  RISE = 1120,
-  REOSC = 2894,
-  RESISTANCE = 356,
-  SOLANA = 501,
-  STACKS = 5757,
-  STELLAR = 148,
-  SHYFT = 7341,
-  TEZOS = 1729,
-  THUNDERCORE = 1001,
-  TOMO = 889,
-  TON = 607,
-  TRON = 195,
-  UBIQ = 108,
-  VECHAIN = 818,
-  WANCHAIN = 5718350,
-  WAVES = 5741564,
-  ZCASH = 133,
-  ZCLASSIC = 147,
-  ZCOIN = 136,
-  ZEN = 121,
-  ZILLIQA = 313,
-  SUI = 784,
-}
