--- libs/ledgerjs/packages/types-cryptoassets/tsconfig.build.json
@@ -1,19 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "customConditions": [],
-    "types": []
-  },
-  "exclude": [
-    "**/*.test.ts",
-    "**/*.test.tsx",
-    "**/*.spec.ts",
-    "**/*.spec.tsx",
-    "**/__tests__/**/*",
-    "**/tests/**/*",
-    "**/*.test.js",
-    "**/*.test.jsx",
-    "**/*.spec.js",
-    "**/*.spec.jsx"
-  ]
-}
