+++ libs/ledgerjs/packages/types-live/src/portfolio.ts
@@ -1,5 +1,5 @@
 import type { AccountLike, AccountLikeArray, GranularityId } from "./account";
-import type { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
+import type { CryptoCurrency, TokenCurrency } from "./currency";
 
 /**
  *
