+++ libs/ledgerjs/packages/types-live/src/swap.ts
@@ -1,4 +1,4 @@
-import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
+import type { CryptoOrTokenCurrency } from "./currency";
 import { BigNumber } from "bignumber.js";
 
 /**
