+++ package.json
@@ -171,7 +171,6 @@
     "live-dmk-shared": "pnpm --filter live-dmk-shared",
     "live-dmk-mobile": "pnpm --filter live-dmk-mobile",
     "live-dmk-desktop": "pnpm --filter live-dmk-desktop",
-    "ljs:cryptoassets": "pnpm --filter cryptoassets",
     "ljs:devices": "pnpm --filter devices",
     "ljs:errors": "pnpm --filter errors",
     "ljs:hw-app:coverage": "nx run-many -t coverage -p \"@ledgerhq/hw-app-*\" --parallel=2",
@@ -205,7 +204,6 @@
     "ljs:hw-transport-node-speculos-http": "pnpm --filter hw-transport-node-speculos-http",
     "ljs:hw-transport-vault": "pnpm --filter hw-transport-vault",
     "ljs:logs": "pnpm --filter logs",
-    "ljs:types-cryptoassets": "pnpm --filter types-cryptoassets",
     "ljs:types-devices": "pnpm --filter types-devices",
     "ljs:types-live": "pnpm --filter types-live",
     "service:cal": "pnpm --filter ledger-cal-service",
